requirements.txt
keras==3.6.0
matplotlib==3.9.2
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
scikit_learn==1.5.2
streamlit==1.39.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
